Added settings in nexus.settings.yml to use the default favicon but a custom logo.
Drupal requires the logo and favicon to be in .svg and .ico formats respectively. Thus the files should be name logo.svg & favicon.ico.
Currently nexus theme has favicon in .ico format but the logo is in .png format. This requires an override to use logo.png as custom logo untill the time a logo in .svg format is created.